Add Multiple buttons and cases
added the nextComputer and previousComputer button.
added the listeners to those buttons.
added tooltips telling people those buttons aren't developed fully.
added cases for the buttons in the actionPerformed method.
added a preformed cases for the kill method in the actionPreformed
method, not sure how to select and grab that information.