Update GUI  to use killtask method with a click.
added logic to verify that they got processes.
Added logic to display processes and add the processes to a list.
Added logic to transalate the clicks to the list.
Added two global variables to project;
one to project to verify that getprocess button was clicked before
trying to select a process to kill.
one to give the killprocess method the process ID to kill.