change vector to JList: GUIComponent
for GUIComponent:
Added JList and set up JList to replace all highlighter, Vector and
MouseListener methods.
for ProcessComponent:
deleted the {4} because it wasn't needed
added a break because we are guaranteed that the first match is the
Process ID.
Deleted the unnecessary if statement as i put that logic in the
GUIComponent.