Built the nextComputer buttons functionality
added a JTextArea to store the computer name on the programs GUI; this
accomplishes showing the user which computer they are currently viewing
the processes on.
added a variable to store the hostID which is the IP v4 of the current
computer.
added a little bit of logic to make sure that the process didn't fail.